Prototype code for Sinusoid Generator to examine those INFOs.
To obtain those INFOs from input SIGNAL (AUDIO), Renpoo made the 1st.
launcher batch script.

Now we can see picked-up AMPLITUDEs and FREQUENCIEs, so are able to
customise this script to get the other INFOs like PHASEs, and also
enough to compute DELTA selection whether or not we have to connect the
SINE wave to the OTHER.